feat: create method to get the triples of the field + validations on that field
Refs: #SFB-8
